[REFACTOR] calendar와 event 도메인 분리 #25
