+++ src/main/java/com/indayvidual/server/domain/calendar/service/CalendarQueryServiceImpl.java
@@ -1,10 +1,10 @@
 package com.indayvidual.server.domain.calendar.service;
 
-import com.indayvidual.server.domain.calendar.converter.EventConverter;
 import com.indayvidual.server.domain.calendar.dto.response.GetDayEventResponseDto;
 import com.indayvidual.server.domain.calendar.dto.response.GetMonthlyCalendarResponseDto;
-import com.indayvidual.server.domain.calendar.entity.Event;
-import com.indayvidual.server.domain.calendar.repository.EventRepository;
+import com.indayvidual.server.domain.event.converter.EventConverter;
+import com.indayvidual.server.domain.event.entity.Event;
+import com.indayvidual.server.domain.event.repository.EventRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -19,21 +19,9 @@ import java.util.stream.Collectors;
 @Service
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
-public class EventQueryServiceImpl implements EventQueryService {
+public class CalendarQueryServiceImpl implements CalendarQueryService {
 
     private final EventRepository eventRepository;
-    private final EventConverter eventConverter;
-
-    @Override
-    public Event findByIdAndUserId(Long eventId, Long userId) {
-        return eventRepository.findByIdAndUserId(eventId, userId)
-                .orElseThrow(() -> new RuntimeException("해당 일정을 찾을 수 없습니다."));
-    }
-
-    @Override
-    public boolean existsByIdAndUserId(Long eventId, Long userId) {
-        return eventRepository.existsByIdAndUserId(eventId, userId);
-    }
 
     @Override
     public List<GetMonthlyCalendarResponseDto> getMonthlyCalendar(int year, int month, Long userId) {
@@ -59,14 +47,5 @@ public class EventQueryServiceImpl implements EventQueryService {
             currentDate = currentDate.plusDays(1);
         }
         return result;
-    }
-
-    @Override
-    public List<GetDayEventResponseDto> getDayEvents(LocalDate date, Long userId) {
-        List<Event> events = eventRepository.findByUserIdAndEventDateOrderByStartTimeAsc(userId, date);
-
-        return events.stream()
-                .map(eventConverter::toGetDayEventResponse)
-                .collect(Collectors.toList());
     }
 }
