--- src/main/java/com/indayvidual/server/domain/calendar/service/EventCommandService.java
@@ -1,11 +1,0 @@
-package com.indayvidual.server.domain.calendar.service;
-
-import com.indayvidual.server.domain.calendar.dto.request.CreateEventRequestDto;
-import com.indayvidual.server.domain.calendar.dto.request.UpdateEventRequestDto;
-import com.indayvidual.server.domain.calendar.entity.Event;
-
-public interface EventCommandService {
-    Event createEvent(CreateEventRequestDto request, Long userId);
-    Event updateEvent(Long eventId, UpdateEventRequestDto request, Long userId);
-    void deleteEvent(Long eventId, Long userId);
-}
