+++ src/main/java/com/indayvidual/server/domain/event/dto/request/UpdateEventRequestDto.java
@@ -1,14 +1,11 @@
-package com.indayvidual.server.domain.calendar.dto.request;
+package com.indayvidual.server.domain.event.dto.request;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import io.swagger.v3.oas.annotations.media.Schema;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Pattern;
 import jakarta.validation.constraints.Size;
 import lombok.*;
 
-import java.time.LocalDate;
 import java.time.LocalTime;
 
 @Getter
