+++ src/main/java/com/indayvidual/server/domain/event/dto/response/CreateEventResponseDto.java
@@ -1,4 +1,4 @@
-package com.indayvidual.server.domain.calendar.dto.response;
+package com.indayvidual.server.domain.event.dto.response;
 
 import lombok.*;
