+++ src/main/java/com/indayvidual/server/domain/event/repository/EventRepository.java
@@ -1,6 +1,6 @@
-package com.indayvidual.server.domain.calendar.repository;
+package com.indayvidual.server.domain.event.repository;
 
-import com.indayvidual.server.domain.calendar.entity.Event;
+import com.indayvidual.server.domain.event.entity.Event;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
