+++ src/main/java/com/indayvidual/server/domain/event/service/EventQueryService.java
@@ -1,8 +1,8 @@
-package com.indayvidual.server.domain.calendar.service;
+package com.indayvidual.server.domain.event.service;
 
 import com.indayvidual.server.domain.calendar.dto.response.GetDayEventResponseDto;
 import com.indayvidual.server.domain.calendar.dto.response.GetMonthlyCalendarResponseDto;
-import com.indayvidual.server.domain.calendar.entity.Event;
+import com.indayvidual.server.domain.event.entity.Event;
 
 import java.time.LocalDate;
 import java.util.List;
@@ -10,7 +10,5 @@ import java.util.List;
 public interface EventQueryService {
     Event findByIdAndUserId(Long eventId, Long userId);
     boolean existsByIdAndUserId(Long eventId, Long userId);
-
-    List<GetMonthlyCalendarResponseDto> getMonthlyCalendar(int year, int month, Long userId);
     List<GetDayEventResponseDto> getDayEvents(LocalDate date, Long userId);
 }
